fixed numpy dtype to be compatible with numpy 2

See merge request cms-analysis/general/HiggsDNA!576

+++ higgs_dna/systematics/event_weight_systematics.py
@@ -913,7 +913,7 @@ def bTagShapeSF(events, weights, ShapeSF_name, is_correction=True, year="2017", 
         )
         # Multiply the scale factore of all jets in a even
 
-        sf = ak.values_astype(dummy_sf, np.float)
+        sf = ak.values_astype(dummy_sf, np.float32)
         sf_central = ak.prod(
             ak.unflatten(_sf_central, counts),
             axis=1
